Tests/LibWeb/Text/input/css/child-change-no-op-invalidation-counters.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    .plain {
        color: rgb(0, 0, 0);
    }
</style>
<div id="parent" class="plain"></div>
<script>
    function settleAndReset(triggerElement) {
        getComputedStyle(triggerElement).color;
        internals.resetStyleInvalidationCounters();
    }

    function verify(label, expected) {
        const invalidations = internals.getStyleInvalidationCounters().styleInvalidations;
        if (invalidations === expected)
            println(`PASS: ${label} (${invalidations} invalidations)`);
        else
            println(`FAIL: ${label} (expected ${expected}, got ${invalidations})`);
    }

    test(() => {
        const parent = document.getElementById("parent");

        settleAndReset(parent);
        const child = document.createElement("span");
        parent.appendChild(child);
        getComputedStyle(parent).color;
        verify("appending a child to a parent without observable child-state pseudos is a no-op", 0);

        settleAndReset(parent);
        parent.removeChild(child);
        getComputedStyle(parent).color;
        verify("removing the only child from a parent without observable child-state pseudos is a no-op", 0);

        settleAndReset(parent);
        const fragment = document.createDocumentFragment();
        fragment.appendChild(document.createElement("span"));
        fragment.appendChild(document.createElement("span"));
        parent.appendChild(fragment);
        getComputedStyle(parent).color;
        verify("appending a fragment to a parent without observable child-state pseudos is a no-op", 0);
    });
</script>
